+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- New version information is automatically added after line 8 -->
 
-## Version 1.4.47 - tbd
+## Version 1.4.47 - 2020-10-08
 
 ### Fixed
 
+- Respect 'Content-ID' in HTTP request headers
 - Update on peer dependencies
+- Update README on OData V2 Adapter for CAP Java 
 - Update README on mission statement
 
 ## Version 1.4.46 - 2020-09-29